All Dao classes implemented.

+++ Sprint 2/BusReservationSystem/src/main/java/com/cg/busreservationsystem/dao/BookingDao.java
@@ -1,5 +1,6 @@
 package com.cg.busreservationsystem.dao;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.cg.busreservationsystem.dto.Booking;
@@ -8,9 +9,9 @@ import com.cg.busreservationsystem.dto.Passenger;;
 public interface BookingDao {
 	
 	public Booking saveBooking(Booking booking);
-	public Integer removeBooking(Integer bookingId);
+	public Integer removeBooking(BigInteger bookingId);
 	public List<Booking> findAllBookings();
-	public Booking findBookingById(Integer bookingId);
+	public Booking findBookingById(BigInteger bookingId);
 	public Passenger savePassenger(Passenger passenger);
 	public List<Passenger> findAllPassengers();
 	public Passenger findPassengerByName(String pname);
